+++ docs/src/constants/routes.js
@@ -68,7 +68,7 @@ export const CHECKBOX = '/components/checkbox';
 export const SNACKBAR = '/components/snackbar';
 export const FORM = '/components/form';
 export const FORM_LABEL = '/components/form-label';
-export const VALIDATION = '/components/validation';
+export const FORM_VALIDATION = '/components/form-validation';
 export const CARD = '/components/card';
 export const CHIP = '/components/chip';
 export const BADGE = '/components/badge';
